+++ main.js
@@ -6,13 +6,17 @@ import neighborhoodHandler from './modules/neighborhoodHandler.js';
 
 (function() {
   if (ls.isAvailable()) {
-    if (!ls.get('NEIGHBORHOODS')) {
+    // Delete old data
+    if (ls.get('NEIGHBORHOODS')) {
+      ls.remove('NEIGHBORHOODS');
+    }
+    if (!ls.get('NEIGHBORHOODSv2')) {
       console.log(
-        'localstorage.NEIGHBORHOODS does not yet exist, but soon should.'
+        'localstorage.NEIGHBORHOODSv2 does not yet exist, but soon should.'
       );
       xhrHandler(true);
     } else {
-      console.log('Awesome, localstorage.NEIGHBORHOODS exists!');
+      console.log('Awesome, localstorage.NEIGHBORHOODSv2 exists!');
       neighborhoodHandler();
     }
   } else {